- put all achievement management together on one page - added being able to see who has the achievement from the manage view - added being able to delete people's achievement from the achievement - made sure you dont always go back to the list after changing an achievement
