--- resources/views/achievement/give.scss
@@ -1,62 +1,0 @@
-.achievement div {
-  float: right;
-  width: 50%;
-  padding: 10px;
-  padding-bottom: 0;
-}
-
-.achievement .achievement-icon {
-  padding:0;
-  text-align:center;
-  max-width:calc(50% - 20px);
-  height:calc(100% - 20px);
-  top:50%;
-  position: absolute;
-  transform:translate(0, -50%);
-}
-
-.list-group {
-  margin-bottom:0;
-}
-
-.achievement-icon img {
-  height: 100px;
-  max-width: 100%;
-  top:50%;
-  left:50%;
-  position: absolute;
-  transform:translate(-50%, -50%);
-}
-
-.achievement {
-  margin-top:5px;
-  overflow: hidden;
-  word-wrap: break-word;
-  border-width: 5px;
-  position: relative;
-  height:120px;
-}
-
-#user-select div {
-  padding: 0;
-}
-
-.COMMON {
-  border-color: #DDDDDD;
-}
-
-.UNCOMMON {
-  border-color: #1E90FF;
-}
-
-.RARE {
-  border-color: #9932CC;
-}
-
-.EPIC {
-  border-color: #333333;
-}
-
-.LEGENDARY {
-  border-color: #C1FF00;
-}
